feat(site): ca advisor demos
